Extract param format converting logic out of cudnn saveables.
This is step 1 for revamp cudnn rnn unittests.

Previously, each CudnnXXXSaveable object takes care of
* Inserting save-restore logic
* Proper tensor stitching to convert between cudnn and tf rnn weight formats.

Now the 2nd part is taken out to separate classes (a.k.a CudnnParamsFormatConverterXXX). This is to later help the effort to add more high quality unittest for current cudnn and later cudnn feature set.

PiperOrigin-RevId: 220575235